requirements.txt
# Core ML / NLP
torch>=2.0.0
transformers==4.44.0
sentence-transformers==3.2.0

# Vector database
chromadb>=0.4.22

# Text preprocessing
nltk>=3.8.1
pymorphy3>=2.0.0           # Russian lemmatization (Python 3.11+ compatible)

# Document loaders
PyPDF2>=3.0.1
python-docx>=1.0.0

# LLM backends
requests>=2.31.0           # For Ollama HTTP API
openai>=1.10.0             # Optional: OpenAI API

# Telegram bot
python-telegram-bot>=21.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
python-dotenv>=1.0.0
